bot/requirements.txt
amqp==2.6.0
asgiref==3.2.7
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
Django==3.0.7
django-celery-beat==2.0.0
django-emoji-picker==0.0.6
django-json-widget==1.0.1
django-timezone-field==4.0
djangorestframework==3.11.0
future==0.18.2
kombu==4.6.10
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-telegram-bot==12.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
vine==1.3.0
wcwidth==0.2.4
gunicorn==20.0.4
Pillow==7.2.0
requests==2.24.0
